DISCUSSION-23: Allow to chose the storage location by "application"
This commit starts a big refactoring of Discussion to integrate the
concept of applicationHint: the idea is to be able to manipulate
discussion with a dedicated applicationHint so that dedicated
applications can be built on top of it. Right now this hint is only used
for choosing the location where to store the discussions, but we could
imagine other usecases later with this hint.