DISCUSSION-37: Discussion read right should be allowed by default
This commit changes the Read right to allow it by default and introduces
new API to be able to chose to allow or deny rights. It also modifies
MessageStream code to create discussions so that no rights is set when a
public discussion is created so that anyone can see it.