DISCUSSION-42: Upload of attachment is not properly supported
Provides a dedicated API to allow creating a message from a request,
without using the discussion handler. The idea behind that work is that
in some cases we want to be able to create discussion message withing
another activity (in my case, I want to use that when performing a
change request review). And we don't necessarily want to perform two
different requests (one for the activity and one for the message
creation). So being able to create the message directly from the request
would be helpful.